scraper/requirements.txt
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
playwright>=1.43.0
python-dotenv>=1.0.0
